Test Reading Update
Changed logging information for the reading of a score file to be debug
level, instead of info level.

Also fixed the reading of names and time to be converted to strings for
comparisons to the file entries.

Added additional info after the processing.

Commented out choice to test find_matches.read_moment(). This never
worked and this tester is best for the test of
process_frames.read_frame().