Fixed python 3 incompatabilities.
Fixed issues in find_matches.py in the logging of bytes data from
ffmpeg. subprocess handles this differently in python 3, causing errors.

Also None < 0 now throws and error, instead of True. Made sure not to
rely on that behavior.

Also changed the executing integreter to python3 on find_matches.py.