Impelmented the finish command.
The finish command is supposed to take a folder of entries, pick the
files with green tags, and add the intro, with customized text, to the
video. The current version just creates the customized text. It still
needs to be worked on so that the text wraps instead of going all the
way across frame. Also, it should be generated at the wanted resolution
and automatically prepended to the video.