Implement Data log.
A full data log that can be set by --data-log and is the cvs list of all
of the read values from frames, along with the frame numbers.

This serves to both write out the read frames, but also to cache the
values in the video if the process should be canceled or crash while
processing.