Cleaned up spaceraid finish operation.
Adjusted some of the comments to fix typos, mistakes, and make it a
little more clear.

Also made data-log part of parse(), instead of global, as it should be.

Removed run from documentation, it is still there though (for now).

Corrected the ffmpeg location from
/usr/local/Cellar/ffmpeg/3.4.1/bin/ffmpeg to /usr/local/bin/ffmpeg.
Also, create the output folder if it does not already exists.

Also, for the text on screen, now os.path.splitext() is used now instead
of str.rstrip() to prevent cutting of extra character.

Now, the python process only sleeps after reading all of ffmpeg to make
sure ffmpeg is not throttled.

Also, a logging error is now recorded if ffmpeg fails, an earlier
oversight.